prev register pages: initialized
