+++ src/main/frontend/src/components/views/PrevRegisterPage/MyPrevRegisterCourseList.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function MyPrevRegisterCourseList() {
+  return (
+    <div>
+      <div>MyPrevRegisterCourseList</div>
+    </div>
+  );
+}
+
+export default MyPrevRegisterCourseList;
