+++ src/main/frontend/src/components/views/PrevRegisterPage/PrevRegisterFilter.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function PrevRegisterFilter() {
+  return (
+    <div>PrevRegisterFilter</div>
+  )
+}
+
+export default PrevRegisterFilter
