+++ src/main/frontend/src/components/views/PrevRegisterPage/PrevRegisterItem.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function PrevRegisterItem() {
+  return <div>PrevRegisterItem</div>;
+}
+
+export default PrevRegisterItem;
