+++ src/main/frontend/src/components/views/PrevRegisterPage/PrevRegisterList.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+function PrevRegisterList() {
+  return (
+    <div>PrevRegisterList</div>
+  )
+}
+
+export default PrevRegisterList
